[Fix] DSV4 attention: int64 overflow, save_for_backward None, CSR bound, remove dead kernels
sparse_mla_dsv4_train.py (kernel):
- Cast query_idx to int64 to prevent pointer arithmetic overflow for
  long sequences.
- Cast slot to int64 before kv_stride_n multiplication to prevent
  int32 wrap on large KV pools.

sparse_mla_dsv4_train.py (wrapper):
- Replace attn_sink=None with empty tensor in ctx.save_for_backward;
  None is not a valid argument to save_for_backward. Backward detects
  no-sink via attn_sink.numel() == 0.
- Filter _build_inverted_topk to reject indices >= num_kv in addition
  to < 0, matching the gather kernel's invalid-slot behavior.

dsv4_indexer.py (kernel):
- Remove _indexer_fwd_kernel, _indexer_bwd_dq_dw_kernel, and
  _indexer_bwd_dk_kernel: these were never called by any wrapper in
  this repo or in Lumen; the public indexer ops use PyTorch for
  scoring. Removing avoids dead-code confusion and unverified paths.

+++ aiter/ops/triton/_triton_kernels/attention/sparse_mla_dsv4_train.py
@@ -87,7 +87,7 @@ def _sparse_mla_fwd_kernel(
     BLOCK_H: tl.constexpr,
     BLOCK_K: tl.constexpr,
 ):
-    query_idx = tl.program_id(0)
+    query_idx = tl.program_id(0).to(tl.int64)
     pid_h = tl.program_id(1)
 
     head_offsets = pid_h * BLOCK_H + tl.arange(0, BLOCK_H)
@@ -115,7 +115,7 @@ def _sparse_mla_fwd_kernel(
         in_range = k_pos < topk
         slot = tl.load(
             indices_ptr + query_idx * idx_stride_t + k_pos, mask=in_range, other=-1
-        )
+        ).to(tl.int64)
         valid = in_range & (slot >= 0) & (slot < num_kv)
 
         kv = tl.load(
